Move projects to separate django app
