+++ backend/pillars/models.py
@@ -6,7 +6,7 @@ class Pillar(models.Model):
         "auth.User", on_delete=models.CASCADE, related_name="pillars"
     )
     project = models.ForeignKey(
-        "game_concept.Project",
+        "projects.Project",
         on_delete=models.SET_NULL,
         null=True,
         blank=True,
