+++ backend/pillars/views_pillar.py
@@ -10,10 +10,10 @@ from rest_framework.request import Request
 from rest_framework.serializers import BaseSerializer
 from rest_framework.viewsets import ModelViewSet, ViewSet
 
-from game_concept.utils import get_current_project
 from llm import LLMOrchestrator
 from llm.types import LLMRequest
 from llm.view_utils import get_model_id
+from projects.utils import get_current_project
 
 from .models import Pillar
 from .serializers import PillarSerializer
