+++ backend/projects/services.py
@@ -15,18 +15,19 @@ from pxcharts.models import (
 )
 from pxnodes.models import PxComponent, PxComponentDefinition, PxNode
 
-from .models import GameConcept, Project
+from game_concept.models import GameConcept
+from .models import Project
 
 
 def clone_project(
-    *,
-    source_project: Project,
-    user: User,
-    name: str,
-    include_concept: bool,
-    include_pillars: bool,
-    include_charts: bool,
-    include_nodes: bool,
+        *,
+        source_project: Project,
+        user: User,
+        name: str,
+        include_concept: bool,
+        include_pillars: bool,
+        include_charts: bool,
+        include_nodes: bool,
 ) -> Project:
     new_project = Project.objects.create(
         user=user,
@@ -83,12 +84,12 @@ def clone_project(
 
 
 def _clone_nodes(
-    *,
-    source_project: Project,
-    new_project: Project,
-    user: User,
-    def_map: dict[str, PxComponentDefinition],
-    node_map: dict[str, PxNode],
+        *,
+        source_project: Project,
+        new_project: Project,
+        user: User,
+        def_map: dict[str, PxComponentDefinition],
+        node_map: dict[str, PxNode],
 ) -> None:
     definitions = PxComponentDefinition.objects.filter(project=source_project)
     for definition in definitions:
@@ -134,12 +135,12 @@ def _clone_nodes(
 
 
 def _clone_charts(
-    *,
-    source_project: Project,
-    new_project: Project,
-    user: User,
-    include_nodes: bool,
-    node_map: dict[str, PxNode],
+        *,
+        source_project: Project,
+        new_project: Project,
+        user: User,
+        include_nodes: bool,
+        node_map: dict[str, PxNode],
 ) -> None:
     charts = PxChart.objects.filter(project=source_project)
     for chart in charts:
@@ -186,7 +187,7 @@ def _clone_charts(
 
 
 def _clone_container_layout(
-    container: PxChartContainer, new_container: PxChartContainer
+        container: PxChartContainer, new_container: PxChartContainer
 ) -> None:
     if not hasattr(container, "layout") or not container.layout:
         return
