+++ backend/projects/urls.py
@@ -7,9 +7,9 @@ from typing import List, Union
 from django.urls import URLPattern, URLResolver
 from rest_framework.routers import DefaultRouter
 
-from .views import GameConceptViewSet
+from .views import ProjectViewSet
 
 router = DefaultRouter()
-router.register(r"", GameConceptViewSet, basename="game-concept")
+router.register(r"", ProjectViewSet, basename="project")
 
 urlpatterns: List[Union[URLPattern, URLResolver]] = router.urls
