+++ backend/pxnodes/views.py
@@ -8,7 +8,8 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from game_concept.utils import get_current_game_concept, get_current_project
+from game_concept.utils import get_current_game_concept
+from projects.utils import get_current_project
 from pillars.models import Pillar
 from pxcharts.models import PxChart
 from pxnodes.llm.context.artifacts import ArtifactInventory
